requirements.txt
amqp==5.3.1
annotated-types==0.7.0
attrs==23.2.0
billiard==4.2.1
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
db-dtypes==1.3.1
et-xmlfile==1.1.0
flower==2.0.1
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.68.0
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
humanize==4.11.0
idna==3.7
kombu==5.4.2
lxml==5.2.2
mysql-connector-python==9.0.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandas-gbq==0.24.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==18.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydata-google-auth==1.8.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.21.0
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.33
sqlmodel==0.0.22
tornado==6.4.2
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
wsproto==1.2.0
